backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
openai
httpx
pydantic
pydantic-settings
stripe
python-jose[cryptography]
pyjwt[crypto]
passlib[bcrypt]
slowapi
psycopg2-binary
email-validator
fastapi-mail
jinja2
pytest
pytest-asyncio
pytest-cov
